Ai/requirements.txt
torch
transformers
ultralytics
Pillow
requests
